+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsefunction.py
@@ -507,9 +507,9 @@ class EvaluatorResponseFunction(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-14T09:44:44.987Z"
+    created: Optional[str] = "2026-07-14T13:06:04.309Z"
 
-    updated: Optional[str] = "2026-07-14T09:44:44.987Z"
+    updated: Optional[str] = "2026-07-14T13:06:04.309Z"
 
     updated_by_id: OptionalNullable[str] = UNSET