+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KXG07WM4VJ3GDWXTR180X9XF"
+        "01KXGBS9P8RE5N3FTTNBFVANG1"
     )
     r"""The unique identifier of the data source"""