+++ packages/orq-rc/RELEASES.md
@@ -9429,3 +9429,13 @@ Based on:
 - [python v4.12.0-rc.21] packages/orq-rc
 ### Releases
 - [PyPI v4.12.0-rc.21] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.21 - packages/orq-rc
+
+## 2026-07-14 13:09:17
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.1 (2.918.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.12.0-rc.22] packages/orq-rc
+### Releases
+- [PyPI v4.12.0-rc.22] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.22 - packages/orq-rc